Adding file browsing, refactoring some code and implementing the ideas from the feedback and applying the MVC pattern
