+++ fileManagment/Model/TypeEnum.swift
@@ -1,0 +1,13 @@
+//
+//  TypeEnum.swift
+//  fileManagment
+//
+//  Created by Rayan Taj on 27/11/2021.
+//
+
+import Foundation
+
+enum TypeCreate: Int {
+    case file = 0
+    case folder = 1
+}
